% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios_por_ugrhi.R
\docType{data}
\name{municipios_por_ugrhi}
\alias{municipios_por_ugrhi}
\title{Municípios por UGRHI}
\format{
\subsection{\code{municipios_por_ugrhi}}{

Uma tibble com 645 observações e 3 colunas:
\describe{
\item{municipio}{Nome do município}
\item{id_municipio}{Código IBGE do município}
\item{n_ugrhi}{Número da UGRHI}
...
}
}
}
\source{
\url{https://sigrh.sp.gov.br/municipios}
}
\usage{
municipios_por_ugrhi
}
\description{
Apresenta dados sobre os municípios que compõem cada Unidade de Gerenciamento
de Recursos Hídricos (UGRHI).
Dados extraídos do Portal SigRH-SP, arrumados e disponibilizados para uso.
Dados extraídos usando raspagem de dados, em 26/11/2023.
}
\keyword{datasets}
